+++ M6502/HarteTest_6502/cycle_t.cpp
@@ -43,7 +43,7 @@ cycle_t::cycle_t(simdjson::dom::array input) noexcept
 #ifdef USE_RAPIDJSON_JSON
 
 cycle_t::cycle_t(const rapidjson::Value& input)
-: m_address((uint8_t)input[0].GetInt64()),
+: m_address((uint16_t)input[0].GetInt64()),
   m_value((uint8_t)input[1].GetInt64()),
   m_action(to_action(input[2].GetString())) {
     assert(input.Size() == 3);
